vdin: fix vdin coverity [1/1]
PD#SWPL-168347

Problem:
vdin need fix coverity problem

Solution:
fix it

Verify:
T5M

+++ drivers/media/vin/tvin/tvin_global.h
@@ -587,7 +587,7 @@ struct tvin_sig_property_s {
 	u8 dolby_vision;/*is signal dolby version 1:vsif 2:emp */
 	bool low_latency;/*is low latency dolby mode*/
 	u8 up_sample_en;/* 420+2ppc 420+4ppc need enable to 1 */
-	u8 fps;
+	u32 fps;
 	unsigned int skip_vf_num;/*skip pre vframe num*/
 	struct tvin_latency_s latency;
 	struct tvin_fmm_s filmmaker;
